Fixed index buffer not working

+++ src/Mesh.h
@@ -30,14 +30,14 @@ private:
 public:
     // Create a new mesh with a set of vertices, indices and textures. The mesh will not be built
     // until build() is called.
-    Mesh(const std::string& pipelineName, std::vector<Vertex> vertices, std::vector<unsigned int> indices, std::vector<Texture> textures);
+    Mesh(const std::string& pipelineName, std::vector<Vertex> vertices, std::vector<unsigned short> indices, std::vector<Texture> textures);
 
     // Create a new empty mesh, make sure you call rebuild() to place vertices, indices and textures within
     // the mesh at a later time.
     Mesh(const std::string& pipelineName);
 
     // Rebuilds the mesh with a new set of vertices, indices and textures
-    void rebuild(std::vector<Vertex> vertices, std::vector<unsigned int> indices, std::vector<Texture> textures, RenderableData input);
+    void rebuild(std::vector<Vertex> vertices, std::vector<unsigned short> indices, std::vector<Texture> textures, RenderableData input);
 
     void build(RenderableData input) override;
     void render(vk::CommandBuffer &commandBuffer, GraphicsPipeline &pipeline) override;
@@ -48,6 +48,6 @@ public:
     [[nodiscard]] bool isBuilt() const { return _built; }
 
     std::vector<Vertex> Vertices;
-    std::vector<unsigned int> Indices;
+    std::vector<unsigned short> Indices;
     std::vector<Texture> Textures;
 };
