+++ src/core/Model.cpp
@@ -45,7 +45,7 @@ void Model::processNode(aiNode *node, const aiScene *scene) {
 
 Mesh Model::processMesh(aiMesh *mesh, const aiScene *scene) {
     std::vector<Vertex> vertices;
-    std::vector<unsigned int> indices;
+    std::vector<unsigned short> indices;
     std::vector<Texture> textures;
 
     for (unsigned int i = 0; i < mesh->mNumVertices; i++) {
